Refactoring of test shell
