+++ TestShell/TestShell.vcxproj
@@ -130,17 +130,17 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="CommandFactory.cpp" />
     <ClCompile Include="CustomException.cpp" />
     <ClCompile Include="Logger.cpp" />
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="Runner.cpp" />
+    <ClCompile Include="ShellCommand.cpp" />
     <ClCompile Include="SsdDriver.cpp" />
     <ClCompile Include="TCManager.cpp" />
     <ClCompile Include="TestShell.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ISSD.h" />
-    <ClInclude Include="Logger.h" />
     <ClInclude Include="TCResult.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
